requirements.txt
absl-py==2.3.0
aiosmtplib==3.0.2
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
fastapi-mail==1.5.0
flatbuffers==25.2.10
fonttools==4.58.4
generativeai==0.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.62.3
h11==0.16.0
httplib2==0.22.0
idna==3.10
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
plyer==2.1.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
scipy==1.16.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==11.0.3
wsproto==1.2.0
numba==0.61.2
llvmlite==0.44.0
psutil==6.1.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.70
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langsmith==0.4.8
faiss-cpu==1.11.0.post1
pypdf==5.8.0
google-auth-oauthlib==1.2.0
google-auth==2.40.3
requests-oauthlib==1.3.1